Added bit to budget.
In particular, I added that we have some funding support from other
sources, so that it is clear that we're 100% go, and that this is a
reasonable request to support the endeavor.

I also added Danny Barnes (who could be removed), because he will be a
full partner in this project this summer. I don't know 100% who else
might join full-time, so he is the only one listed.

Added Jan, added affiliations.